+++ src/app/project/view/project.view.component.ts
@@ -1,0 +1,72 @@
+import { Component } from '@angular/core';
+import { MdCard } from '@angular2-material/card/card';
+import { MdButton } from '@angular2-material/button/button';
+import { MdToolbar } from '@angular2-material/toolbar/toolbar';
+import { MdIcon, MdIconRegistry } from '@angular2-material/icon/icon';
+import { MD_GRID_LIST_DIRECTIVES } from '@angular2-material/grid-list/grid-list';
+import { ProjectModel, ProjectService } from '../../shared/project/index';
+import { ROUTER_DIRECTIVES, Router, ActivatedRoute } from '@angular/router';
+
+@Component({
+  moduleId: module.id,
+  templateUrl: './project.view.component.html',
+  styleUrls: ['../../../assets/css/skeleton.css', './project.view.component.css'],
+  directives: [ROUTER_DIRECTIVES, MD_GRID_LIST_DIRECTIVES, MdCard, MdButton, MdIcon, MdToolbar],
+  providers: [MdIconRegistry, ProjectService]
+})
+
+export class ProjectViewComponent {
+
+  project: ProjectModel = new ProjectModel()
+
+  /**
+   * Dependency Injection: route (for reading params later)
+   */
+  constructor(
+    private router: Router,
+    private route: ActivatedRoute,
+    private projectService: ProjectService
+  ){}
+
+  /**
+   * Initialization Event Handler, used to parse route params
+   * like `id` in project/:id/edit)
+   */
+  ngOnInit() {
+    this.route.params
+      .map(params => params['id'])
+      .subscribe((id) => {
+        console.log('View Project by ID from route params:', id)
+        if (id) {
+          this.projectService.getProject(id)
+          .subscribe(
+            data => {
+              this.setProject(data)
+            },
+            err => console.error(err),
+            () => {}
+          )
+        }
+      })
+  }
+
+  /**
+   * Project loading handler
+   * @param {data} Loaded project data
+   */
+  setProject(data){
+    this.project = data;
+  }
+
+  /**
+   * Remove this project
+   * @param {project} Project being viewed
+   */
+  private deleteProject(project: ProjectModel) {
+    // Delete from DB
+    this.projectService.deleteProject(project)
+
+    this.router.navigate(['/projects'])
+    return false;
+  }
+}