+++ src/main.ts
@@ -6,12 +6,14 @@ import { provideRouter } from '@angular/router';
 
 import { AboutComponent } from './app/about/index';
 import { LandingComponent } from './app/landing/index';
+
 import { LeaderEditComponent } from './app/leader/edit/index';
 import { LeaderListComponent } from './app/leader/list/index';
 import { LeaderViewComponent } from './app/leader/view/index';
 
 import { ProjectEditComponent } from './app/project/edit/index';
 import { ProjectListComponent } from './app/project/list/index';
+import { ProjectViewComponent } from './app/project/view/index';
 
 if (environment.production) {
   enableProdMode();
@@ -20,15 +22,19 @@ if (environment.production) {
 bootstrap(AppComponent, [
   HTTP_PROVIDERS,
   provideRouter([
+
     { index: true, component: LandingComponent },
-    { path: '/about', component: AboutComponent },
 
-    { path: '/leader/:id', component: LeaderViewComponent },
-    { path: '/leader/:id/edit', component: LeaderEditComponent },
-    { path: '/leaders', component: LeaderListComponent },
     { path: '/add-leader', component: LeaderEditComponent },
+    { path: '/leader/:id', component: LeaderViewComponent },
+    { path: '/leaders', component: LeaderListComponent },
+    { path: '/leader/:id/edit', component: LeaderEditComponent },
 
+    { path: '/add-project', component: ProjectEditComponent },
+    { path: '/project/:id', component: ProjectViewComponent },
     { path: '/projects', component: ProjectListComponent },
-    { path: '/add-project', component: ProjectEditComponent }
+    { path: '/project/:id/edit', component: ProjectEditComponent },
+
+    { path: '/about', component: AboutComponent }
   ])
 ]);