+++ docs/changelog.md
@@ -28,9 +28,17 @@
 12.08.2016
 
 Лайв-застосунок запрацював тут: https://live-europe.rhcloud.com/
-
 Тепер він на адекватному NodeJS, як і раніше має ССЛ, додалася працююча БД
 
+14.08.2016
+
+* Project CRUD / DB integration:
+ * Added ability to Create Project
+ * Ability to View Project
+ * Ability to Edit Project
+ * Ability to Delete Project
+* Project List, Project View, Project Editor
+* Implemented dynamic Landing page incl. leaders and projects count
 
 
 ## Почитати: